logout mekanisme update

+++ src/components/organism/crystalharvester/auth/reset/index.tsx
@@ -29,10 +29,10 @@ export default function CrystalHarvesterReset() {
         const onSubmit=()=>{
             const {username,password,password_confirm}=formdata
             const {playerName}=storage
-            if(username === playerName && password === password_confirm){
+            if(username === playerName && password === password_confirm && password.length > 6 && password_confirm.length > 6){
                 localStorage.setItem('password',password)
                 localStorage.setItem('token','true')
-                // Kuki__initialize()
+                Kuki__initialize()
                 navigate('/loading/crystal&login')
             }
         }
@@ -53,24 +53,24 @@ export default function CrystalHarvesterReset() {
             <div className="input-wrapper">
                 <label htmlFor="password" className='d-flex gap-1 align-items-center'>
                     password
-                    <Icon variant={formdata.password_confirm === formdata.password?'info':'danger'} icon={'check2-circle'} name={''}/>
+                    <Icon variant={formdata.password_confirm === formdata.password && formdata.password_confirm.length > 6 && formdata.password.length > 6 ?'info':'danger'} icon={'check2-circle'} name={''}/>
                 </label>
                 <input type="password" id='password' placeholder='password' onChange={onChange} />
             </div>
             <div className="input-wrapper">
                 <label htmlFor="password_confirm" className='d-flex gap-1 align-items-center'>
                     confirm password
-                    <Icon variant={formdata.password_confirm === formdata.password?'info':'danger'} icon={'check2-circle'} name={''}/>
+                    <Icon variant={formdata.password_confirm === formdata.password && formdata.password_confirm.length > 6 && formdata.password.length > 6 ?'info':'danger'} icon={'check2-circle'} name={''}/>
                 </label>
                 <input type="password" id='password_confirm' placeholder='confirm password' onChange={onChange} />
             </div>
             <button className='auth-btn' onClick={onSubmit}>reset</button>
             <hr/>
             <div className="d-flex gap-3">
-                <NavLink to={'loading/crystal&login'} className='link-harvester'>
+                <NavLink to={'/loading/crystal&register'} className='link-harvester'>
                     register
                 </NavLink>
-                <NavLink to={'loading/crystal&login'} className='link-harvester'>
+                <NavLink to={'/loading/crystal&login'} className='link-harvester'>
                     login
                 </NavLink>
             </div>
